refactor(design): clean up  email  documentation structure
